+++ src/components/AddBook.js
@@ -1,19 +1,43 @@
-import React from 'react';
-import PropTypes from 'prop-types';
+import React, { useState } from 'react';
+import { useDispatch } from 'react-redux';
+import { addBook } from '../redux/books/books';
 
-const AddBook = ({ submitBook, bookDetail }) => (
-  <>
-    <form id="form" onSubmit={submitBook}>
-      <input type="text" id="book-title" name="title" className="book" placeholder="Book Name..." onChange={bookDetail} required />
-      <input type="text" id="book-author" name="author" className="book" placeholder="Book Author..." onChange={bookDetail} required />
-      <input type="submit" />
-    </form>
-  </>
-);
+const AddBook = () => {
+  const [title, setTitle] = useState('');
+  const [author, setAuthor] = useState('');
+  const dispatch = useDispatch();
+  const submitBook = (e) => {
+    e.preventDefault();
+    const form = e.target;
+    const book = {
+      id: new Date().getTime().toString(),
+      title,
+      author,
+    };
+    dispatch(addBook(book));
+    setTitle('');
+    setAuthor('');
+    form.reset();
+  };
 
-AddBook.propTypes = {
-  submitBook: PropTypes.func.isRequired,
-  bookDetail: PropTypes.func.isRequired,
+  const bookDetail = (e) => {
+    const { value } = e.target;
+    if (e.target.id === 'book-title') {
+      setTitle(value);
+    } else {
+      setAuthor(value);
+    }
+  };
+
+  return (
+    <>
+      <form id="form" onSubmit={submitBook}>
+        <input type="text" id="book-title" name="title" className="book" placeholder="Book Name..." onChange={bookDetail} required />
+        <input type="text" id="book-author" name="author" className="book" placeholder="Book Author..." onChange={bookDetail} required />
+        <input type="submit" />
+      </form>
+    </>
+  );
 };
 
 export default AddBook;
